requirements.txt
asttokens==2.0.5
backcall==0.2.0
chardet==4.0.0
colorama==0.4.4
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
ipykernel==6.12.1
ipython==8.2.0
jedi==0.18.1
jupyter-client==7.2.1
jupyter-core==4.9.2
matplotlib-inline==0.1.3
mysql-connector-python==8.0.28
nest-asyncio==1.5.5
numpy==1.22.3
openpyxl==3.0.9
packaging==21.3
pandas==1.4.2
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.29
protobuf==3.20.0
psutil==5.9.0
pure-eval==0.2.2
Pygments==2.11.2
pyodbc @ file:///C:/Users/INNO/Downloads/pyodbc-4.0.32-cp310-cp310-win_amd64.whl
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pywin32==303
pyzmq==22.3.0
six==1.16.0
stack-data==0.2.0
tornado==6.1
traitlets==5.1.1
wcwidth==0.2.5
xlrd==2.0.1
